Issue #152 - Rename widgetPosition in ProjectWidgetPositionDTO and use this class to send projectWidgetPosition through ProjectWidgetDto instance
